requirements-dev.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.5.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
requests==2.32.3
setuptools==75.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
whitenoise==6.8.2
